requirements.txt
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
idna==2.7
inflection==0.3.1
kiwisolver==1.0.1
lxml==4.2.5
matplotlib==3.0.0
more-itertools==4.3.0
numpy==1.15.2
pandas==0.23.4
pandas-datareader==0.7.0
pep8==1.7.1
pycodestyle==2.4.0
pyparsing==2.2.1
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
Quandl==3.4.2
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
urllib3==1.23
wrapt==1.10.11
